+++ engine/src/maths/vmaths.c
@@ -1,0 +1,15 @@
+#include "vmaths.h"
+
+#include <math.h>
+
+f64 vsin(f64 value) {
+    return sin(value);
+}
+
+f64 vcos(f64 value) {
+    return cos(value);
+}
+
+f64 vtan(f64 value) {
+    return tan(value);
+}