+++ engine/src/renderer/open-gl/opengl-backend.c
@@ -1,20 +1,22 @@
 #include "opengl-backend.h"
 #include "opengl-shaders.h"
+#include "opengl-buffers.h"
+#include "opengl-arrays.h"
 
 
-
-b8 BackendInitializeGLFW(openGLContext* context, i32 width, i32 height) {
+b8 BackendInitializeGLFW(OpenGLContext* context, i32 width, i32 height) {
     //Initialize GLFW
     glfwInit();
     glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
     glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 3);
     glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
 
+
     //Set window context
-    context->window = glfwCreateWindow(800, 600, "OpenGL Testing", NULL, NULL);
+    context->window = glfwCreateWindow(800, 600, "Victoria Engine", NULL, NULL);
     if (context->window == NULL)
     {
-        VFATAL("Failed to create window!");
+        VFATAL("Failed to create GLFW window.");
         glfwTerminate();
         return FALSE;
     }
@@ -23,102 +25,104 @@ b8 BackendInitializeGLFW(openGLContext* context, i32 width, i32 height) {
 
     //Load GLAD
     if (!gladLoadGLLoader((GLADloadproc)glfwGetProcAddress)) {
-        VFATAL("Failed to load GLAD!");
+        VFATAL("Failed to load GLAD.");
         return FALSE;
     }
-    
-    //TODO: Separate functions
 
-    //Initialize shader program
-    GLShaderInitializeShaderProgram(&context->shaderProgram);
-
-    //TODO: Temporary code. Find a way to use tables to have more than one.
-    ShaderObject fragShader;
-    //DO THIS!!!!!
-    fragShader.type = FRAGMENT_SHADER;
-    fragShader.filepath = "C:/Programs/C-and-C++/Victoria-Engine/assets/test_shader.frag.glsl";
-    //ShaderObject vertShader;
-
-    //TODO: Look for materials and set up their respective fragment and vertex shaders
-    for (u32 i = 0; i < 10; i++) {
-        if (i == 1) {
-            RegisterNewShaderObject(&fragShader, &context->shaderProgram);
-        }
-        //if (i == 2) {
-        //    RegisterNewShaderObject(&vertShader, &context->shaderProgram);
-        //}
+    /* Shader program initialization */
+    if (!GLInitializeShaderProgram(&context->shaderProgram)) {
+       VFATAL("Failed to initialise shader program.");
+       return FALSE;
     }
 
+    // Shader setup
+    Shader myShader;
+    myShader.name = "test_shader";
 
-    GLLinkShaderProgram(&context->shaderProgram);
+    //TODO: Uniform lists
 
-    //TODO: Delete shaders
-    //glDeleteShader(vertexShader);
-    //glDeleteShader(fragmentShader);
+    if (!GLRegisterNewShader(&myShader, &context->shaderProgram)) {
+        VFATAL("Shader failed to be initialized.");
+        return FALSE;
+    }
+    glLinkProgram(context->shaderProgram);
+    GLDeleteShaders(&myShader);
 
-    f32 vertices[] = {
-     0.5f, 0.5f, 0.0f,
+    // Buffer object setup
+    float vertices[] = {
+    -0.5f, -0.5f, 0.0f,
      0.5f, -0.5f, 0.0f,
-     -0.5f,  -0.5f, 0.0f,
-     -0.5f, 0.5f, 0.0f
-    };
-    u32 indicies[] = {
-        0, 1, 3, // Triangle 1
-        1, 2, 3  // Triangle 2
-    };
+     0.0f,  0.5f, 0.0f
+    };  
 
+    //VAO creation and binding (functions overshadow the default for now)
+    GLGenerateArrayObject(&context->vertexArrayObj);
+    GLBindVertexArray(context->vertexArrayObj);
 
-    glGenVertexArrays(1, &context->VAO);
-    glGenBuffers(1, &context->VBO);  
-    glGenBuffers(1, &context->EBO);
-    glBindVertexArray(context->VAO);
+    /* VBO initialization and binding */
+    if (!RegisterNewBufferObject(&context->vertexBufferObj, ARRAY_BUF)) {
+        VFATAL("VBO could not be registered.");
+        return FALSE;
+    }
+    glBufferData(ARRAY_BUF, sizeof(vertices), vertices, STATIC_DRAW);
 
-    glBindBuffer(GL_ARRAY_BUFFER, context->VBO);
-    glBufferData(GL_ARRAY_BUFFER, sizeof(vertices), vertices, GL_STATIC_DRAW);
-    
-    //EBO buffer
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, context->EBO);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(indicies), indicies, GL_STATIC_DRAW);
-
+    /* Function info:
+    * First parameter is the layout location for the vertex shader (i.e. the layout (location = 0)) part
+    * Second pararm is the size of the vertex attribute (it's a vec3 so 3)
+    * Third param is the type of data (vec3 is float data)
+    * Fourth param is if the data needs to be normalized (if it's a int or byte)
+    * Fifth param is the stride or the size of each vertex attribute
+    * Sixth param is a void* to the offset of the data (if we have colour data after each vertex it'd be the offset from that)
+    */
     glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 3 * sizeof(f32), (void*)0);
+    // Enable the layout in the location above
     glEnableVertexAttribArray(0);
 
-    glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-    glBindVertexArray(0);
-
-    //TODO: temp wireframe view
-    //glPolygonMode(GL_FRONT_AND_BACK, GL_LINE);
 
     return TRUE;
 }
 
-b8 BackendRenderFrame(openGLContext* context, i32 width, i32 height) {
-    //TODO: Temporary code. Fix later
+b8 BackendRenderFrame(OpenGLContext* context, i32 width, i32 height) {
+    f32 startFrame = glfwGetTime();
+
+    // Input processing
+    ProcessInput(context->window);
+
+    // Create background colour
     glClearColor(0.2f, 0.3f, 0.3f, 1.0f);
     glClear(GL_COLOR_BUFFER_BIT);
 
-
+    //Uniform detection
     glUseProgram(context->shaderProgram);
-
-    glBindVertexArray(context->VAO);
-    //EBO buffer stuff
-    glDrawElements(GL_TRIANGLES, 6, GL_UNSIGNED_INT, 0);
+    f32 greenValue = (vsin(startFrame) / 2.0f) + 0.5f;
+    i32 vertexColourLoc = glGetUniformLocation(context->shaderProgram, "ourColour");
+    glUniform4f(vertexColourLoc, 0.0f, greenValue, 0.0f, 1.0f);
+    
+    // Draw objects
+    //glUseProgram(context->shaderProgram);
+    glBindVertexArray(context->vertexArrayObj);
+    glDrawArrays(GL_TRIANGLES, 0, 3);
 
     glfwSwapBuffers(context->window);
     glfwPollEvents();
+    f32 endFrame = glfwGetTime();
+
+    //Check to see if the window should close
+    context->windowShouldClose = glfwWindowShouldClose(context->window);
+    context->deltaTime = endFrame - startFrame;
     return TRUE;
 }
 
-b8 BackendShutdownRendering(openGLContext* context) {
-    if (glfwWindowShouldClose(context->window)) {
-        VERROR("Window already closing!");
+b8 BackendShutdownRendering(OpenGLContext* context) {
+    // In the case that the application failed to load our window, this happens.
+    if (context->window == 0) {
+        VERROR("Context doesn't exist! Booting out...");
         return FALSE;
     }
+    
     glfwSetWindowShouldClose(context->window, TRUE);
-    glDeleteVertexArrays(1, &context->VAO);
-    glDeleteBuffers(1, &context->VBO);
-    glDeleteBuffers(1, &context->EBO);
+    glDeleteVertexArrays(1, &context->vertexArrayObj);
+    glDeleteBuffers(1, &context->vertexBufferObj);
     glDeleteProgram(context->shaderProgram);
     glfwTerminate();
     VINFO("Backend successfully terminated.");
@@ -128,4 +132,10 @@ b8 BackendShutdownRendering(openGLContext* context) {
 void BackendFramebufferSizeCallback(GLFWwindow* window, i32 width, i32 height) {
     // Typecasting just to be safe
     glViewport(0, 0, (int)width, (int)height);
+}
+
+void ProcessInput(GLFWwindow* window) {
+    if (glfwGetKey(window, GLFW_KEY_ESCAPE) == GLFW_PRESS) {
+        glfwSetWindowShouldClose(window, TRUE);
+    }
 }