+++ engine/src/renderer/open-gl/opengl-backend.h
@@ -2,23 +2,18 @@
 
 #include "defines.h"
 #include "core/logger.h"
+#include "maths/vmaths.h"
 
-#include <glad/glad.h>
-#include <GLFW/glfw3.h>
+#include "opengl-types.inl"
 
-typedef struct openGLContext {
-    GLFWwindow* window;
-    //TODO: move to separate struct
-    u32 shaderProgram;
-    u32 VAO;
-    u32 VBO;
-    u32 EBO;
-} openGLContext;
 
-b8 BackendInitializeGLFW(openGLContext* context, i32 width, i32 height);
+b8 BackendInitializeGLFW(OpenGLContext* context, i32 width, i32 height);
 
-b8 BackendRenderFrame(openGLContext* context, i32 width, i32 height);
+b8 BackendRenderFrame(OpenGLContext* context, i32 width, i32 height);
 
-b8 BackendShutdownRendering(openGLContext* context);
+b8 BackendShutdownRendering(OpenGLContext* context);
 
 void BackendFramebufferSizeCallback(GLFWwindow* window, i32 width, i32 height);
+
+//TODO: Better input mapping, this is temporary.
+void ProcessInput(GLFWwindow* window);